backend/requirements.txt
Flask==3.1.1
flask_cors==6.0.0
Flask_JWT_Extended==4.7.1
flask_restx==1.3.0
openpyxl==3.1.5
pandas==2.3.0
PyJWT==2.10.1
PyJWT==2.10.1
PyMySQL==1.1.1
Werkzeug==3.1.3
